Manages subscribers
- Adds paper_trail gem that gives us the advantage of tracking
changes to our models, for auditing or versioning, and also we'll be
able to see how a model looked at any stage in its lifecycle, revert it
to any version, or restore it after it has been destroyed.
- Implements the undo destroy action when deleting subscribers.
- Uses Angular event propagation for undo destroy action, for that reason
a new directive is created to handle the dom manipulation.
- Adds destroy action to email list through the implementation of ngDialog
in order to inform the users about the operation they are about to perform
and they will be asked to confirm by typing a special text before to be
able to delete that resource.
- Implements infinite scroll to list subscribers, this feature involves
kaminari to control de pagination.